Refactor Webpack part of XDL
