+++ packages/expo-cli/package.json
@@ -74,7 +74,7 @@
     "@expo/package-manager": "^0.0.1",
     "@expo/simple-spinner": "1.0.2",
     "@expo/spawn-async": "1.5.0",
-    "@expo/xdl": "57.2.1",
+    "@expo/xdl": "57.3.1",
     "@types/cli-table": "^0.3.0",
     "@types/untildify": "^3.0.0",
     "ansi-regex": "^4.1.0",
