+++ packages/xdl/src/Webpack/WebpackOptions.ts
@@ -1,0 +1,70 @@
+import * as ProjectSettings from '../ProjectSettings';
+import { BundlingOptions, CLIWebOptions, WebEnvironment } from './Webpack.types';
+import { isInfoEnabled } from './WebpackEnv';
+
+export function validateBoolOption(name: string, value: unknown, defaultValue: boolean): boolean {
+  if (typeof value === 'undefined') {
+    value = defaultValue;
+  }
+
+  if (typeof value !== 'boolean') {
+    throw new Error(`Invalid webpack option: '${name}'. Option must be a boolean.`);
+  }
+
+  return value;
+}
+
+export function transformCLIOptions(options: CLIWebOptions): BundlingOptions {
+  // Transform the CLI flags into more explicit values
+  return {
+    ...options,
+    isImageEditingEnabled: options.pwa,
+  };
+}
+
+async function applyOptionsToProjectSettingsAsync(
+  projectRoot: string,
+  options: BundlingOptions
+): Promise<ProjectSettings.Settings> {
+  let newSettings: Partial<ProjectSettings.Settings> = {};
+  // Change settings before reading them
+  if (typeof options.https === 'boolean') {
+    newSettings.https = options.https;
+  }
+
+  if (Object.keys(newSettings).length) {
+    await ProjectSettings.setAsync(projectRoot, newSettings);
+  }
+
+  return await ProjectSettings.readAsync(projectRoot);
+}
+
+export async function getWebpackConfigEnvFromBundlingOptionsAsync(
+  projectRoot: string,
+  options: BundlingOptions
+): Promise<WebEnvironment> {
+  let { dev, https } = await applyOptionsToProjectSettingsAsync(projectRoot, options);
+
+  const mode = typeof options.mode === 'string' ? options.mode : dev ? 'development' : 'production';
+
+  const isImageEditingEnabled = validateBoolOption(
+    'isImageEditingEnabled',
+    options.isImageEditingEnabled,
+    true
+  );
+  const isDebugInfoEnabled = validateBoolOption(
+    'isDebugInfoEnabled',
+    options.isDebugInfoEnabled,
+    isInfoEnabled()
+  );
+
+  return {
+    projectRoot,
+    pwa: isImageEditingEnabled,
+    isImageEditingEnabled,
+    mode,
+    https,
+    info: isDebugInfoEnabled,
+    ...(options.webpackEnv || {}),
+  };
+}
